Just replace the old PatchMapWithMove
