+++ src/Data/Patch.hs
@@ -9,6 +9,9 @@ module Data.Patch
   , module X
   ) where
 
+import Data.Map.Monoidal (MonoidalMap)
+import Data.Semigroup (Semigroup (..), (<>))
+
 import Data.Patch.Class as X
 import Data.Patch.DMap as X hiding (getDeletions)
 import Data.Patch.DMapWithMove as X (PatchDMapWithMove, const2PatchDMapWithMoveWith, mapPatchDMapWithMove,
@@ -20,12 +23,6 @@ import Data.Patch.Map as X
 import Data.Patch.MapWithMove as X (PatchMapWithMove, patchMapWithMoveNewElements,
                                       patchMapWithMoveNewElementsMap, unPatchMapWithMove,
                                       unsafePatchMapWithMove)
-import Reflex.Patch.MapWithMove.New as X (PatchMapWithMove2, patchMapWithMove2NewElements,
-                                      patchMapWithMove2NewElementsMap, unPatchMapWithMove2,
-                                      unsafePatchMapWithMove2)
-
-import Data.Map.Monoidal (MonoidalMap)
-import Data.Semigroup (Semigroup (..), (<>))
 
 -- | A 'Group' is a 'Monoid' where every element has an inverse.
 class (Semigroup q, Monoid q) => Group q where
